Attempt to use hardware timer pulse generation for better color stability
Currently, OE pulses are generated purely in software with a timer
update interrupt. This is susceptible to considerable jitter especially
since the interrupt must be below the system RTOS priority to allow it
to signal the display task when to swap buffers. This technique
still uses update interrupts, but drives the OE pin from a timer compare
channel, which should have much less imprecision.