Adjust dupm queue timeout to 0 for requests from serial
This prevents a deadlock of the form:
    - dupm queue fills
    - dupm waits for packet
    - serial gets a DATA_TEMP message
    - serial waits for 5 seconds for empty space in dupm
    - dupm wait timeout expires
    - stm sends response for dupm packet
    - serial 5 second timeout expires
    - serial dumps out dupm wait response

Since the dupm retries things, this usually repeats a few times as well,
often cascading until no traffic can be exchanged.

Really, there shouldn't be any reason to block for these requests: the
STM retries both of the packets mentioned on its own.