Add fix for memory leak in sdk code
When reconnecting to a WPA2-Enterprise network, the system can decide
to cache on a reconnection attempt, in which case the EAP sm is not
cleaned up and roughly 7k is leaked (well, it's referenced by a global
so will not accumulate over multiple attempts, but still) as a result.

This adds a few hacks to ensure the SDK code is told about key
negotiation succeeding regardless - this is complicated by the fact that
the WPA2 code in libnet80211.a (as decompiled) only checks that state
and deinitializes from rx_eapol, i.e. after a packet is received (or
from connection timeout), so if the state is never set to FINISHED
during a packet reception, it never deinits the SM.

The caching code in esp_wpa2.c responds to a START call with an error if
caching is 'detected', but the underlying code from upstream
wpa_supplicant can restart the EAP sm if the server rejects a cached
PMK. This patch seems to make the SDK code clean up properly, although
in practice this could probably be fixed by lazy-allocating the state on
the first actual EAP packet received, but that's a larger change.