Relax restrictions around ranged timespecs in summarizer
improves situations like

.##......... ##

where you have a long uncertain tail and then a short gap, and
then more precip; the summarizer instead takes center() in all cases
which sounds dumb because it'll be measuring as if

.#######     ##

(taking from the center of the uncertain region)

which gives really weird summaries.

Now computes minute_diff as the small gap, which is what we want.